Added Power Bar and Game Instance
Created basic power bar UI, implemented the power mechanic for player/BP_Scaleable, and added a Game Instance BP to pass the invert y setting between menu and the map. Also added a new exclude to the third party folder that is excluded from VC. Did a music test in the folder. The song composition is in public domain and this particular version is CC 4.0 but unsure if we can distribute this on our repo with regards to the license, so excluding to be better safe than sorry. May not stay in final version regardless.