New Prop Type
Adding new BP_Environmental which can be used to place objects in the world that will stop scaling objects but not scale themselves.
New setup for affected blueprints to cast against these object types to set and unset the ability to grow.
Updating BP_Button to have the trigger volume as the outliner parent.
Updating test level with these changes.