Elevator Updates
Updating game instance to store the current elevator stop and the latest (furthest) one we have achieved.
Elevator buttons and stops! You can use the button and stops together with a linked elevator on the button to set up a bunch of floors.
Adding a floor environmental asset so I can easily test elevator floors that still block scaleables.
Updating interactable to also set a bool for children to know the player is standing inside the trigger for things like the elevator. For some reason overlap won't let me call to the parent overlap.
Updating elevator BP with the changes as it needs to act as a pseudo elevator button but only to send you to the latest checkpoint.
Updating bouncy castle with an elevator system. The second floor is totally unreachable right now unless you do play from here.