Saving and Elevators!
Added saving and loading to the game instance.
Added an elevator the front of the tower as well as various checkpoints on the way up.
Refactored how elevator status is being written as my variables were messy as hell on what we need to save.
Hooked up save events to elevator buttons and applying setting in both option menus.
Loading is held by the level and set up for LVL_Tower.